add TextView Property Composable
